Final Commit

+++ Assets/HealEffect.cs
@@ -6,8 +6,8 @@ public class HealEffect :Effect
 {
     public override void GameUpdate()
     {
-        int currentHealth = transform.gameObject.GetComponent<CharacterStats>().health;
-        int maxHealth = transform.gameObject.GetComponent<CharacterStats>().maxHealth;
+        float currentHealth = transform.gameObject.GetComponent<CharacterStats>().health;
+        float maxHealth = transform.gameObject.GetComponent<CharacterStats>().maxHealth;
         transform.gameObject.GetComponent<CharacterStats>().health += 5;
         if (currentHealth > maxHealth)
         {
